迭代更新

+++ scheduler-cluster/src/main/java/com/bh/spider/scheduler/cluster/worker/WorkerSchedulerComponentAssistant.java
@@ -7,6 +7,7 @@ import com.bh.spider.scheduler.Config;
 import com.bh.spider.scheduler.cluster.ClusterNode;
 import com.bh.spider.scheduler.cluster.component.ComponentOperationEntry;
 import com.bh.spider.scheduler.cluster.consistent.operation.Entry;
+import com.bh.spider.scheduler.cluster.consistent.operation.Operation;
 import com.bh.spider.scheduler.component.ComponentCoreFactory;
 import com.bh.spider.scheduler.component.ComponentRepository;
 import com.bh.spider.scheduler.context.Context;
@@ -15,6 +16,8 @@ import com.bh.spider.scheduler.event.CollectionParams;
 import com.bh.spider.scheduler.event.Command;
 import com.bh.spider.scheduler.event.CommandHandler;
 import com.bh.spider.scheduler.watch.Watch;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -24,9 +27,13 @@ import java.util.List;
 import java.util.concurrent.Callable;
 
 public class WorkerSchedulerComponentAssistant implements Assistant {
+    private final static Logger logger = LoggerFactory.getLogger(WorkerSchedulerComponentAssistant.class);
 
     private WorkerScheduler scheduler;
+
+
     private int loadClassTimeout;
+
     private final ClusterNode node;
 
     private ComponentCoreFactory factory;
@@ -36,10 +43,12 @@ public class WorkerSchedulerComponentAssistant implements Assistant {
 
     private long localCommittedIndex;
 
+    private long version;
+
 
     public WorkerSchedulerComponentAssistant(Config cfg, WorkerScheduler scheduler) throws IOException {
-
         this.scheduler = scheduler;
+
         this.factory = new ComponentCoreFactory(cfg);
 
         this.loadClassTimeout = Integer.parseInt(cfg.get(Config.INIT_LOAD_CLASS_TIMEOUT));
@@ -67,26 +76,39 @@ public class WorkerSchedulerComponentAssistant implements Assistant {
 
 
     @CommandHandler
-    public void WRITE_OPERATION_ENTRIES(Context ctx, @CollectionParams(collectionType = List.class, argumentTypes = {Entry.class}) List<Entry> entries) throws IOException {
+    public void WRITE_OPERATION_ENTRIES(Context ctx, @CollectionParams(collectionType = List.class, argumentTypes = {Entry.class}) List<Entry> entries, long version) throws IOException {
 
+        if (version != this.version) return;
 
+        long committedIndex = localCommittedIndex;
         //更新component metadata
         for (Entry entry : entries) {
-            ComponentOperationEntry coe = new ComponentOperationEntry(entry);
+            if (entry.index() < committedIndex) {
+                logger.error("过时的日志条目:{},当前commitIndex:{}", entry.index(), committedIndex);
+                continue;
+            }
+            if (entry.action() == Operation.SNAP) {
+                this.factory.apply(entry.data());
+                committedIndex = entry.index();
+            } else {
+                if (committedIndex != entry.index() - 1) continue;
 
-            if (ComponentOperationEntry.ADD.equals(coe.operation())) {
-                ComponentRepository repository = factory.proxy(coe.type());
+                ComponentOperationEntry coe = new ComponentOperationEntry(entry);
 
-                //如果是JAR包，则立刻进行下载,否则，仅暂时写入一个空文件
-                if (coe.type() == Component.Type.JAR) {
-                    Command cmd = new Command(ctx, CommandCode.WORKER_GET_COMPONENT.name(), coe.name());
-                    ctx.write(cmd);
-                }
-                repository.save(new byte[0], coe.name(), null, true, true);
-            } else if (ComponentOperationEntry.DELETE.equals(coe.operation())) {
-                ComponentRepository repository = factory.proxy(coe.name());
-                if (repository != null) {
-                    repository.delete(coe.name());
+                if (ComponentOperationEntry.ADD.equals(coe.operation())) {
+                    ComponentRepository repository = factory.proxy(coe.type());
+
+                    //如果是JAR包，则立刻进行下载,否则，仅暂时写入一个空文件
+                    if (coe.type() == Component.Type.JAR) {
+                        Command cmd = new Command(ctx, CommandCode.WORKER_GET_COMPONENT.name(), coe.name());
+                        ctx.write(cmd);
+                    }
+                    repository.save(new byte[0], coe.name(), null, true, true);
+                } else if (ComponentOperationEntry.DELETE.equals(coe.operation())) {
+                    ComponentRepository repository = factory.proxy(coe.name());
+                    if (repository != null) {
+                        repository.delete(coe.name());
+                    }
                 }
             }
         }
@@ -94,15 +116,14 @@ public class WorkerSchedulerComponentAssistant implements Assistant {
 
         //写入数据
 
-        this.localCommittedIndex = writeIndex(committedIndexPath,0);
+        this.localCommittedIndex = writeIndex(committedIndexPath, committedIndex);
     }
 
-    @CommandHandler
-    public void CHECK_COMPONENT_OPERATION_COMMITTED_INDEX_HANDLER(Context ctx, long committedIndex) {
-        if (localCommittedIndex < committedIndex) {
-            Command cmd = new Command(ctx, CommandCode.GET_COMPONENT_OPERATION_INDEX.name(), localCommittedIndex);
-            ctx.write(cmd);
-        }
+
+    @CommandHandler(cron = "*/10 * * * * ?")
+    public void CHECK_COMPONENT_OPERATION_COMMITTED_INDEX_HANDLER() {
+        Command cmd = new Command(null, CommandCode.CHECK_COMPONENT_OPERATION_COMMITTED_INDEX.name(), localCommittedIndex);
+        scheduler.communicator().random().write(cmd);
     }
